+++ sdk/cosmos/azure-cosmos/azure/cosmos/documents.py
@@ -412,6 +412,8 @@ class _QueryFeature:
     OrderBy: Literal["OrderBy"] = "OrderBy"
     Top: Literal["Top"] = "Top"
     NonStreamingOrderBy: Literal["NonStreamingOrderBy"] = "NonStreamingOrderBy"
+    HybridSearch: Literal["HybridSearch"] = "HybridSearch"
+    CountIf: Literal["CountIf"] = "CountIf"
 
 
 class _DistinctType: